# book-network-api/src/main/java/com/alibou/book/feedback/FeedbackResponse.java
package com.alibou.book.feedback;

import lombok.*;

@Getter
@Setter
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class FeedbackResponse {
    private Double note;
    private String comment;
    private boolean ownFeedback;

}